requirements.txt
requests>=2.28.0
urllib3>=1.26.0
keboola.component>=1.0.0
python-dotenv>=1.0.0
